rails g synergy:install works
