+++ lib/spree/synergy_configuration.rb
@@ -18,4 +18,10 @@ module Spree
     preference :director, :string
     preference :accountant, :string
   end
+
+  AppConfiguration.class_eval do
+    preference :store_email, :string, :default => 'test@site.ru'
+    preference :store_phone, :string, :default => '+7 (495) 1234567'
+    preference :store_address, :string, :default => 'Somewhere st. Nowhere'
+  end
 end
