+++ lib/synergy.rb
@@ -1,0 +1,2 @@
+require 'spree_core'
+require 'spree/synergy'
